Split Random::Formatter from SecureRandom [Feature #18190]
ruby/ruby@8cd5f5a5a776c67715b7ddb42d5a3b6b2957dc21
